requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
python-dotenv==0.19.2
dynaconf==3.1.8
Flask==3.0.3
Flask-SQLAlchemy==3.0.1
greenlet==3.0.3
gunicorn==22.0.0
isapi-wsgi==0.4.2
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
psycopg2==2.9.4
python-dotenv==0.19.2
SQLAlchemy==2.0.29
toml==0.10.2
typing_extensions==4.11.0
waitress==3.0.0
Werkzeug==3.0.2